refactor: reorganize project structure for better readability
- Create organized directory structure:
  - src/core/ for main logger components
  - src/utils/ for utility functions
  - src/__tests__/ for test files
- Split utils.ts into specialized modules:
  - crypto.ts for UUID/request ID generation
  - serialization.ts for safe JSON handling
  - time.ts for timestamp utilities
- Update all import paths to match new structure
- All tests pass, TypeScript compiles, linting clean
